gprof
I have not made graph as it assumes that the user has gprof2dot.py file.
In order to get profiling data, the Box2d window has to be closed once it is executed. I am assuming that the user will close the window himself to get profiling data for desired length of time.